admin_dashboard/agents/interview_agent/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
RealtimeSTT
ollama
kokoro-onnx
sounddevice
numpy
python-multipart
